Split out jsconfig data conversion to extra class
